reader

+++ include/fdf.h
@@ -6,7 +6,7 @@
 /*   By: aeclipso <aeclipso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/29 11:00:30 by aeclipso          #+#    #+#             */
-/*   Updated: 2020/10/29 18:20:01 by aeclipso         ###   ########.fr       */
+/*   Updated: 2020/11/07 17:12:27 by aeclipso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@
 
 typedef struct	s_map
 {
-	size_t		height;
-	size_t		width;
+	int		height;
+	int		width;
 	int			**map;
 }				t_map;
