+++ main.c
@@ -6,7 +6,7 @@
 /*   By: aeclipso <aeclipso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/29 10:47:46 by aeclipso          #+#    #+#             */
-/*   Updated: 2020/10/29 18:24:04 by aeclipso         ###   ########.fr       */
+/*   Updated: 2020/11/07 17:13:38 by aeclipso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ void		usage(void)
 
 void			delete_map(t_map *map)
 {
-	size_t		i;
+	int			i;
 	int			**del;
 
 	i = 0;
